requirements.txt
Telethon~=1.39.0
python-dotenv~=1.1.0